Read config from home folder
